moved failure class in utils package
